refactor(payments-list-core): add get-payments logic back to the component
